Move kernel panic implementation to the kfmt package
